Refactor leader-elect related code
